adds the `:fetch_from_base` option for input parameters ...
Parameters can now be given the option `:fetch_from_base`:

```ruby
expects do
  string :customer_id, fetch_from_base: true
end
```

If that parameter is not provided when the endpoint is called, we
check for it at `Apidiesel::Api#config[:customer_id]` or
`Apidiesel::Api#customer_id` and use their values if present.

This makes handling easier in cases where lots of different
endpoints require one and the same fixed parameter - you just provide
it once when initialising the `Api` base.