introduces multiple endpoint configurations through "actions" ...
Actions allow you to configure different sets of behaviour into a
single endpoint - for example, one for GET and one for POST requests.

```ruby
module Endpoints
  class User < Apidiesel::Api
    url path: "/user"

    action(:get) do
      expects do
        integer :id
      end

      responds_with do
        integer :id
        string :firstname
        string :lastname
      end
    end

    action(:post) do
      expects do
        integer :id
        string :firstname
        string :lastname
      end
    end
  end
end
```

Under the hood, an action is an anonymous subclass of the parent
class, so actions inherit their parents settings (`url path: "/user"`
in this example).