refactors configuration management through `Apidiesel::Config`
Also reverts on "official" endpoint class inheritance - this is
tricky to fully get right as it is, and should be scrapped.

Instead of juggling with dozens of instance and class instance
variables all over the place and merging them on a case-by-case
basis, we now have `Apidiesel::Config` as a centralized config
store.

Stores can simply be chained, so that each callsite only has to
interact with "its" config store and still get the benefit of
configuration inheritance.