always uses a Processors::Hash for processing Hash-parts of a response, even if no block is given
Previously when no block was given, we would use a Processors::Primitive
instead, and have it simply call `#to_hash` on the received data.

This means we cannot reliably detect Hash processors when building
a root processor, which in turn means that this construct currently
breaks:

  responds_with do
    hash :some_value
  end

because we cannot reliably detect whether to wrap the given response
processors in a Hash or an Array processor.